fix comments requests
